// backend/src/common/shared/enums/user-role.enum.ts
export enum UserRole {
  ADMIN = 'admin',
  EMPLOYEE = 'EMPLOYEE',
  VENDOR = 'vendor',
  CUSTOMER = 'customer',
  DELIVERY = 'delivery',
}

// src/shared/enums/permission.enum.ts
export enum Permission {
  CREATE_ORDER = 'create_order',
  CANCEL_ORDER = 'cancel_order',
  MANAGE_MENU = 'manage_menu',
  APPROVE_VENDOR = 'approve_vendor',
}